+++ extensions/tiktok/src/main/java/app/revanced/extension/tiktok/feedfilter/LikeCountFilter.java
@@ -1,17 +1,18 @@
 package app.revanced.extension.tiktok.feedfilter;
 
-import app.revanced.extension.tiktok.settings.Settings;
 import com.ss.android.ugc.aweme.feed.model.Aweme;
 import com.ss.android.ugc.aweme.feed.model.AwemeStatistics;
 
-import static app.revanced.extension.tiktok.Utils.parseMinMax;
+import app.revanced.extension.tiktok.Utils;
+import app.revanced.extension.tiktok.settings.Settings;
 
 public final class LikeCountFilter implements IFilter {
+
     final long minLike;
     final long maxLike;
 
     LikeCountFilter() {
-        long[] minMax = parseMinMax(Settings.MIN_MAX_LIKES);
+        long[] minMax = Utils.parseMinMax(Settings.MIN_MAX_LIKES);
         minLike = minMax[0];
         maxLike = minMax[1];
     }