+++ extensions/tiktok/src/main/java/app/revanced/extension/tiktok/settings/preference/RangeValuePreference.java
@@ -1,5 +1,6 @@
 package app.revanced.extension.tiktok.settings.preference;
 
+import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -14,11 +15,10 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import app.revanced.extension.shared.settings.StringSetting;
+import app.revanced.extension.tiktok.Utils;
 
 @SuppressWarnings("deprecation")
 public class RangeValuePreference extends DialogPreference {
-    private final Context context;
-
     private String minValue;
 
     private String maxValue;
@@ -29,7 +29,6 @@ public class RangeValuePreference extends DialogPreference {
 
     public RangeValuePreference(Context context, String title, String summary, StringSetting setting) {
         super(context);
-        this.context = context;
         setTitle(title);
         setSummary(summary);
         setKey(setting.key);
@@ -53,41 +52,52 @@ public class RangeValuePreference extends DialogPreference {
         return mValue;
     }
 
+    @SuppressLint("SetTextI18n")
     @Override
     protected View onCreateDialogView() {
         minValue = getValue().split("-")[0];
         maxValue = getValue().split("-")[1];
+
+        Context context = getContext();
+
         LinearLayout dialogView = new LinearLayout(context);
         dialogView.setOrientation(LinearLayout.VERTICAL);
+
+        // Min view
         LinearLayout minView = new LinearLayout(context);
         minView.setOrientation(LinearLayout.HORIZONTAL);
+        dialogView.addView(minView);
+
         TextView min = new TextView(context);
         min.setText("Min: ");
         minView.addView(min);
+
         EditText minEditText = new EditText(context);
         minEditText.setInputType(InputType.TYPE_CLASS_NUMBER);
         minEditText.setText(minValue);
         minView.addView(minEditText);
-        dialogView.addView(minView);
+
+        // Max view
         LinearLayout maxView = new LinearLayout(context);
         maxView.setOrientation(LinearLayout.HORIZONTAL);
+        dialogView.addView(maxView);
+
         TextView max = new TextView(context);
         max.setText("Max: ");
         maxView.addView(max);
+
         EditText maxEditText = new EditText(context);
         maxEditText.setInputType(InputType.TYPE_CLASS_NUMBER);
         maxEditText.setText(maxValue);
         maxView.addView(maxEditText);
-        dialogView.addView(maxView);
+
         minEditText.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-
             }
 
             @Override
             public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-
             }
 
             @Override
@@ -98,12 +108,10 @@ public class RangeValuePreference extends DialogPreference {
         maxEditText.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-
             }
 
             @Override
             public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-
             }
 
             @Override
@@ -111,12 +119,21 @@ public class RangeValuePreference extends DialogPreference {
                 maxValue = editable.toString();
             }
         });
+
         return dialogView;
     }
 
     @Override
+    protected void onBindView(View view) {
+        super.onBindView(view);
+
+        Utils.setTitleAndSummaryColor(getContext(), view);
+    }
+
+    @Override
     protected void onPrepareDialogBuilder(AlertDialog.Builder builder) {
-        builder.setPositiveButton(android.R.string.ok, (dialog, which) -> this.onClick(dialog, DialogInterface.BUTTON_POSITIVE));
+        builder.setPositiveButton(android.R.string.ok, (dialog, which)
+                -> this.onClick(dialog, DialogInterface.BUTTON_POSITIVE));
         builder.setNegativeButton(android.R.string.cancel, null);
     }